In application.properties I changed the spring.jpa.hibernates.ddl-auto=none to =update and reran the application trying to repopulate the DB to correct errors I was having. Then set it back to =none and restarted it, but this did not change anything in the DB as far as I can tell. Keeping it for now and will delete later unless there turns out to be a use for it.
In Customer entity I added 2 imports to lines 4 and 5 for JsonIgnore and JsonIgnoreProperties to prevent recursion.

Line 30 added @JsonIgnoreProperties ({"hibernateLazyInitializer","handler"}) to avoid "proxy noise" by telling Jackson to exclude these internal Hibernate specific fields the serialized fields from the serialized JSON output.

Commented out line 51 and pasted it to line 52 and changed from snake to camelCase for postalCode to maintain consistency and in case it was causing or might cause any errors.

Commented out code for lines 66 and 67 and pasted them to lines 70 and 72 and set optional and nullable = true after checking the MySQL customers table.
Was getting 500 error when testing the frontend on http://localhost:4200 after doing checking out and if any seeded customers have NULL divisions it can't be set to nullable = false and the DB in MySQL shows division_id NULL YES so it is nullable.